HDDS-6387.[Multi-Tenant] Refactor OMMultiTenantManager and OMTenantRequestHelper
