+++ hadoop-ozone/ozone-manager/src/main/java/org/apache/hadoop/ozone/om/OMMultiTenantManager.java
@@ -17,14 +17,9 @@
 package org.apache.hadoop.ozone.om;
 
 import java.io.IOException;
-import java.util.List;
 
 import com.google.common.base.Optional;
-import org.apache.commons.lang3.tuple.Pair;
 import org.apache.hadoop.ozone.om.helpers.TenantUserList;
-import org.apache.hadoop.ozone.om.multitenant.AccessPolicy;
-import org.apache.hadoop.ozone.om.multitenant.AccountNameSpace;
-import org.apache.hadoop.ozone.om.multitenant.BucketNameSpace;
 import org.apache.hadoop.ozone.om.multitenant.Tenant;
 import org.apache.http.auth.BasicUserPrincipal;
 
@@ -68,37 +63,6 @@ public interface OMMultiTenantManager {
    * @return Tenant interface.
    */
   Tenant createTenantAccessInAuthorizer(String tenantID) throws IOException;
-
-
-  /**
-   * Given a TenantID String, Return Tenant Interface. If the Tenant doesn't
-   * exist in the system already, throw Exception.
-   *
-   * @param tenantID
-   * @return Tenant interface.
-   * @throws IOException
-   */
-  Tenant getTenantInfo(String tenantID) throws IOException;
-
-  /**
-   * Given a TenantID String, deactivate the Tenant. If the Tenant has active
-   * users and volumes, tenant gets dectivated. This means
-   * * No new write/modify operations allowed under that tenant.
-   * * No new users can be added.
-   * * All the users of that tenant will not be able to create new
-   *    bucket/keys and carry out any new type of write/update opertations in
-   *    Tenant bucketNamespace or accountNamespace.
-   * * If Tenant has users they will be able to do read and delete operations.
-   * * If the Tenant doesn't have any user or buckets, Tenant will be removed
-   *    from the system.
-   *
-   * * If the Tenant doesn't exist in the system already, throw Exception.
-   *
-   * @param tenantID
-   * @return Tenant interface.
-   * @throws IOException
-   */
-  void deactivateTenant(String tenantID) throws IOException;
 
   /**
    * Given a TenantID, destroys all state associated with that tenant.
@@ -153,44 +117,12 @@ public interface OMMultiTenantManager {
   String getDefaultAccessId(String tenantId, String userPrincipal);
 
   /**
-   * Given a user, return their S3-Secret Key.
-   * @param accessID
-   * @return S3 secret Key
-   */
-  String getUserSecret(String accessID) throws IOException;
-
-  /**
-   * Modify the groups that a user belongs to.
-   * @param accessID
-   * @param groupsAdded
-   * @param groupsRemoved
-   * @throws IOException
-   */
-  void modifyUser(String accessID, List<String> groupsAdded,
-                  List<String> groupsRemoved) throws IOException;
-
-  /**
-   * Given a user, deactivate them. We will need a recon command/job to cleanup
-   * any data owned by this user (ReconMultiTenantManager).
-   * @param accessID
-   */
-  void deactivateUser(String accessID) throws IOException;
-
-  /**
    * Check if a user is a tenant Admin.
    * @param user user name.
    * @param tenantId tenant name.
    * @return
    */
   boolean isTenantAdmin(String user, String tenantId);
-
-  /**
-   * Check if a tenant exists.
-   * @param tenantId tenant name.
-   * @return true if tenant exists, false otherwise.
-   * @throws IOException
-   */
-  boolean tenantExists(String tenantId) throws IOException;
 
   /**
    * List all the user & accessIDs of all users that belong to this Tenant.
@@ -200,15 +132,6 @@ public interface OMMultiTenantManager {
    */
   TenantUserList listUsersInTenant(String tenantID, String prefix)
       throws IOException;
-
-  /**
-   * List all the access IDs of all users that belong to this Tenant.
-   * @param tenantID
-   * @return List of users
-   */
-  List<String> listAllAccessIDs(String tenantID)
-      throws IOException;
-
   /**
    * Given an access ID return its corresponding tenant.
    * @param accessID
@@ -227,91 +150,4 @@ public interface OMMultiTenantManager {
    * Given a user, remove him as admin of the corresponding Tenant.
    */
   void revokeTenantAdmin(String accessID) throws IOException;
-
-  /**
-   * List all the Admin users that belong to this Tenant.
-   * @param tenantID
-   * @return List of users
-   */
-  List<String> listAllTenantAdmin(String tenantID)
-      throws IOException;
-
-  /**
-   * grant given user access to the given BucketNameSpace.
-   * @param accessID
-   * @param bucketNameSpace
-   */
-  void grantAccess(String accessID,
-                   BucketNameSpace bucketNameSpace) throws IOException;
-
-  /**
-   * grant given user access to the given Bucket.
-   * @param accessID
-   * @param bucketNameSpace
-   */
-  void grantBucketAccess(String accessID,
-                   BucketNameSpace bucketNameSpace, String bucketName)
-      throws IOException;
-
-  /**
-   * revoke user access from the given BucketNameSpace.
-   * @param accessID
-   * @param bucketNameSpace
-   */
-  void revokeAccess(String accessID,
-                    BucketNameSpace bucketNameSpace) throws IOException;
-
-  /**
-   * grant given user access to the given AccountNameSpace.
-   * @param accessID
-   * @param accountNameSpace
-   */
-  void grantAccess(String accessID,
-                   AccountNameSpace accountNameSpace) throws IOException;
-
-  /**
-   * revoke user access from the given AccountNameSpace.
-   * @param accessID
-   * @param accountNameSpace
-   */
-  void revokeAccess(String accessID,
-                    AccountNameSpace accountNameSpace) throws IOException;
-
-  /**
-   * Create given policy for the tenant.
-   * @param tenant
-   * @param policy
-   * @return ID of the policy
-   */
-  String createTenantDefaultPolicy(Tenant tenant, AccessPolicy policy)
-      throws IOException;
-
-  /**
-   * Returns default Access policies for a Tenant. Default access policies
-   * are system defined and can not be changed by anyone.
-   * @param tenant
-   * @return list of Default Access policies for a Tenant
-   */
-  List<Pair<String, AccessPolicy>> listDefaultTenantPolicies(Tenant tenant)
-      throws IOException;
-
-  /**
-   * Returns All Access policies for a Tenant. In future we may support
-   * bucket-policies/user-policies to provide cross-tenant accesses.
-   * @param tenant
-   * @return list of Default Access policies for a Tenant
-   */
-  List<Pair<String, AccessPolicy>> listAllTenantPolicies(Tenant tenant)
-      throws IOException;
-
-  /**
-   * Update given policy identified by policyID for the tenant.
-   * @param tenant
-   * @param policyID
-   * @param policy
-   * @return ID of the policy
-   */
-  void updateTenantPolicy(Tenant tenant, String policyID,
-                          AccessPolicy policy) throws IOException;
-
 }
