+++ hadoop-ozone/ozone-manager/src/main/java/org/apache/hadoop/ozone/om/request/s3/tenant/OMTenantAssignAdminRequest.java
@@ -26,6 +26,7 @@ import org.apache.hadoop.hdds.utils.db.cache.CacheValue;
 import org.apache.hadoop.ozone.OzoneConsts;
 import org.apache.hadoop.ozone.audit.OMAction;
 import org.apache.hadoop.ozone.om.OMMetadataManager;
+import org.apache.hadoop.ozone.om.OMMultiTenantManagerImpl;
 import org.apache.hadoop.ozone.om.OzoneManager;
 import org.apache.hadoop.ozone.om.exceptions.OMException;
 import org.apache.hadoop.ozone.om.helpers.OmDBAccessIdInfo;
@@ -47,8 +48,8 @@ import java.util.HashMap;
 import java.util.Map;
 
 import static org.apache.hadoop.ozone.om.lock.OzoneManagerLock.Resource.VOLUME_LOCK;
-import static org.apache.hadoop.ozone.om.request.s3.tenant.OMTenantRequestHelper.checkTenantAdmin;
-import static org.apache.hadoop.ozone.om.request.s3.tenant.OMTenantRequestHelper.checkTenantExistence;
+import static org.apache.hadoop.ozone.om.OMMultiTenantManagerImpl.checkTenantAdmin;
+import static org.apache.hadoop.ozone.om.OMMultiTenantManagerImpl.checkTenantExistence;
 import static org.apache.hadoop.ozone.om.upgrade.OMLayoutFeature.MULTITENANCY_SCHEMA;
 
 /*
@@ -82,7 +83,7 @@ public class OMTenantAssignAdminRequest extends OMClientRequest {
 
     // If tenantId (tenant name) is not provided, infer it from the accessId
     if (StringUtils.isEmpty(tenantId)) {
-      tenantId = OMTenantRequestHelper.getTenantIdFromAccessId(
+      tenantId = OMMultiTenantManagerImpl.getTenantIdFromAccessId(
           ozoneManager.getMetadataManager(), accessId);
       assert (tenantId != null);
     }
@@ -173,7 +174,7 @@ public class OMTenantAssignAdminRequest extends OMClientRequest {
     String volumeName = null;
 
     try {
-      volumeName = OMTenantRequestHelper.getTenantVolumeName(
+      volumeName = OMMultiTenantManagerImpl.getTenantVolumeName(
           omMetadataManager, tenantId);
 
       acquiredVolumeLock = omMetadataManager.getLock().acquireWriteLock(
